+++ README-groovyConsole-MDH.mm
@@ -4,15 +4,15 @@
 <node TEXT="README&#xa;groovyConsole" LOCALIZED_STYLE_REF="AutomaticLayout.level.root" ID="ID_1090958577">
 <hook NAME="accessories/plugins/AutomaticLayout.properties" VALUE="ALL"/>
 <font BOLD="true"/>
-<hook NAME="MapStyle" background="#f5f5dc" zoom="0.909">
-    <properties show_icon_for_attributes="true" edgeColorConfiguration="#808080ff,#ff0000ff,#0000ffff,#00ff00ff,#ff00ffff,#00ffffff,#7c0000ff,#00007cff,#007c00ff,#7c007cff,#007c7cff,#7c7c00ff" show_note_icons="true" fit_to_viewport="false" associatedTemplateLocation="template:/light_sepia_template.mm" mapUsesOwnSaveOptions="true" save_folding="never_save_folding" save_last_visited_node="default" save_modification_times="false"/>
+<hook NAME="MapStyle" background="#f5f5dc" zoom="0.826">
+    <properties show_icon_for_attributes="true" edgeColorConfiguration="#808080ff,#ff0000ff,#0000ffff,#00ff00ff,#ff00ffff,#00ffffff,#7c0000ff,#00007cff,#007c00ff,#7c007cff,#007c7cff,#7c7c00ff" mapUsesOwnSaveOptions="true" save_modification_times="false" save_last_visited_node="default" show_note_icons="true" associatedTemplateLocation="template:/light_sepia_template.mm" save_folding="never_save_folding" fit_to_viewport="false"/>
 
 <map_styles>
 <stylenode LOCALIZED_TEXT="styles.root_node" ID="ID_867972350" STYLE="oval" UNIFORM_SHAPE="true" VGAP_QUANTITY="24 pt">
 <font SIZE="24"/>
 <stylenode LOCALIZED_TEXT="styles.predefined" POSITION="right" STYLE="bubble">
 <stylenode LOCALIZED_TEXT="default" ID="ID_1558627382" ICON_SIZE="12 pt" FORMAT_AS_HYPERLINK="false" COLOR="#2c2b29" BACKGROUND_COLOR="#eedfcc" STYLE="bubble" SHAPE_HORIZONTAL_MARGIN="8 pt" SHAPE_VERTICAL_MARGIN="5 pt" NUMBERED="false" FORMAT="STANDARD_FORMAT" TEXT_ALIGN="DEFAULT" BORDER_WIDTH_LIKE_EDGE="false" BORDER_WIDTH="1.9 px" BORDER_COLOR_LIKE_EDGE="true" BORDER_COLOR="#f0f0f0" BORDER_DASH_LIKE_EDGE="true" BORDER_DASH="SOLID" VGAP_QUANTITY="3 pt" MAX_WIDTH="10 cm" MIN_WIDTH="0 cm">
-<arrowlink SHAPE="CUBIC_CURVE" COLOR="#000000" WIDTH="2" TRANSPARENCY="200" DASH="" FONT_SIZE="9" FONT_FAMILY="SansSerif" DESTINATION="ID_1558627382" STARTINCLINATION="81.75 pt;-10.5 pt;" ENDINCLINATION="81.75 pt;19.5 pt;" STARTARROW="NONE" ENDARROW="DEFAULT"/>
+<arrowlink SHAPE="CUBIC_CURVE" COLOR="#000000" WIDTH="2" TRANSPARENCY="200" DASH="" FONT_SIZE="9" FONT_FAMILY="SansSerif" DESTINATION="ID_1558627382" STARTINCLINATION="81.75 pt;-9.75 pt;" ENDINCLINATION="81.75 pt;19.5 pt;" STARTARROW="NONE" ENDARROW="DEFAULT"/>
 <font NAME="SansSerif" SIZE="9" BOLD="false" STRIKETHROUGH="false" ITALIC="false"/>
 <edge STYLE="bezier" COLOR="#2e3440" WIDTH="3" DASH="SOLID"/>
 <richcontent CONTENT-TYPE="plain/auto" TYPE="DETAILS"/>
@@ -48,7 +48,7 @@
 <stylenode TEXT="MarkdownHelperLink" COLOR="#dbffdb" BACKGROUND_COLOR="#4c4c7f" STYLE="rectangle" BORDER_WIDTH="4 px" BORDER_COLOR_LIKE_EDGE="false" BORDER_COLOR="#009000">
 <icon BUILTIN="emoji-1F517"/>
 </stylenode>
-<stylenode TEXT="MarkdownHelperPreview" COLOR="#333333" BACKGROUND_COLOR="#ffffff">
+<stylenode TEXT="MarkdownHelperPreview" ID="ID_647144206" COLOR="#adbac7" BACKGROUND_COLOR="#22272e">
 <font NAME="Tahoma" SIZE="14"/>
 <hook NAME="NodeCss">pre {
     background-color: #e5e7ff;
@@ -106,14 +106,18 @@ blockquote {
 </stylenode>
 </map_styles>
 </hook>
-<node TEXT="README.MD" STYLE_REF="MarkdownHelperNode" POSITION="right" ID="ID_160804169" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovy_console_2/README.MD">
+<node TEXT="Freeplane_groovyConsole" POSITION="right" ID="ID_108058596" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole">
+<attribute NAME="MarkdownRootFolder" VALUE="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/" OBJECT="java.net.URI|file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/"/>
+<attribute NAME="MDHGithubBranch" VALUE="/main/wip"/>
+<attribute NAME="MDHTargetRootPath" VALUE="https://github.com/EdoFro/Freeplane_groovyConsole"/>
+<node TEXT="README.MD" STYLE_REF="MarkdownHelperNode" ID="ID_160804169" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/README.MD">
 <attribute_layout NAME_WIDTH="138.75 pt" VALUE_WIDTH="42 pt"/>
-<attribute NAME="headersToUnderline" VALUE="0" OBJECT="org.freeplane.features.format.FormattedNumber|0"/>
+<attribute NAME="headersToUnderline" VALUE="3" OBJECT="org.freeplane.features.format.FormattedNumber|3|#0.####"/>
 <attribute NAME="hideFolded" VALUE="false"/>
-<attribute NAME="headerNumbering" VALUE="false"/>
+<attribute NAME="headerNumbering" VALUE="true"/>
 <attribute NAME="topHeadersNumbered" VALUE="false"/>
 <attribute NAME="topHeaderStartingNumber" VALUE="1" OBJECT="org.freeplane.features.format.FormattedNumber|1"/>
-<attribute NAME="fileLinksRelative" VALUE="false"/>
+<attribute NAME="fileLinksRelative" VALUE="true"/>
 <attribute NAME="lineOverHeader" VALUE="true"/>
 <attribute NAME="ignoreHeaderDetails" VALUE="true"/>
 <attribute NAME="ignoreHeaderNotes" VALUE="true"/>
@@ -129,33 +133,338 @@ blockquote {
       = edofro.MarkDownHelper.MDH.document(node)
     </p>
   </body>
-</html>
-</richcontent>
+</html></richcontent>
 <node TEXT="new groovy console for Freeplane" ID="ID_1992959063">
 <node TEXT="A groovy console that runs on groovy 4 and java 17 to replace wikdshell add-on" ID="ID_101985004"/>
-</node>
-</node>
-<node TEXT="README.MD" POSITION="right" ID="ID_181733889" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovy_console_2/README.MD"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/markdown">
+<node TEXT="about the Groovy Console" ID="ID_276201467">
+<node TEXT="The **Groovy Swing Console** allows a user to enter and run **Groovy scripts**.&#xa;It is a standard component from the Groovy language." ID="ID_1511747675"/>
+<node TEXT="It looks like this:" ID="ID_1691832209"/>
+<node TEXT="web Image" STYLE_REF="MarkdownHelperNode" ID="ID_1380829106"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
 <html>
   <head>
     
   </head>
   <body>
     <p>
-      # new groovy console for Freeplane
-    </p>
-    <p>
-      
-    </p>
-    <p>
-      A groovy console that runs on groovy 4 and java 17 to replace wikdshell add-on
-    </p>
-    <p>
-      
+      = edofro.MarkDownHelper.MDH.webImageLink(node)
     </p>
   </body>
 </html>
 </richcontent>
+<node TEXT="Groovy Console" ID="ID_1233743796" LINK="https://docs.groovy-lang.org/latest/html/documentation/assets/img/GroovyConsole.png"/>
+</node>
+<node TEXT="text block" STYLE_REF="MarkdownHelperNode" ID="ID_566564267"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.textBlock(node)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="more info about it can be found in the" ID="ID_1976210177"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      more info about it can be found in the $1
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="web link" STYLE_REF="MarkdownHelperNode" ID="ID_282730740"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.webLink(node)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="Groovy Console official page" ID="ID_486619326" LINK="https://docs.groovy-lang.org/latest/html/documentation/groovy-console.html"/>
+</node>
+</node>
+</node>
+</node>
+<node TEXT="about groovyConsole add-on" ID="ID_1052461647">
+<node TEXT="groovyConsole is a new add-on that opens a Groovy Console and works with Groovy 4 and can be used with the new versions of Freeplane." ID="ID_422303135"/>
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1242531216"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_987299215" LINK="#ID_1941588475"/>
+</node>
+<node TEXT="Additional features:" ID="ID_318297123"/>
+<node TEXT="list" STYLE_REF="MarkdownHelperNode" ID="ID_1192258797"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.list(node)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="It has access to the Freeplane API so it can interact with nodes and maps" ID="ID_210953"/>
+<node TEXT="It can load / save script files to be used with Freeplane" ID="ID_355686780"/>
+<node TEXT="It can &apos;save&apos; scripts as node&apos;s notes or attributes" ID="ID_132684427"/>
+<node TEXT="If you select a node that has a script and then you open the groovy console, the script will be automatically loaded in the console editor" ID="ID_1835156213"/>
+<node TEXT="You can edit the script and &apos;save&apos; them back to the node" ID="ID_678764906"/>
+<node TEXT="You can choose if you want to save the script in its original node or into the currently selected one" ID="ID_589626010">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1272649912"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_1698327657" LINK="#ID_1984441739"/>
+</node>
+</node>
+<node TEXT="You can choose if you want to &apos;save&apos; it as a note or as &apos;script1&apos; attribute" ID="ID_1419130972">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_608855915"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_1544126096" LINK="#ID_1034407721"/>
+</node>
+</node>
+</node>
+<node TEXT="The script stays in the map, so you don&apos;t need to have multiple files for you short scripts" ID="ID_604112988"/>
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1206519521"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_1045513064" LINK="#ID_1693528645"/>
+</node>
+<node TEXT="Additional buttons" ID="ID_484549552">
+<node TEXT="table" STYLE_REF="MarkdownHelperNode" ID="ID_515923454"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.table(node)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="r" ID="ID_826496699">
+<node TEXT="Button" ID="ID_888907803"/>
+<node TEXT="Label" ID="ID_16150301"/>
+<node TEXT="Description" ID="ID_1559268293"/>
+</node>
+<node TEXT="r" ID="ID_1103530420">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1698703557"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_598307815" LINK="#ID_1736357139"/>
+</node>
+<node TEXT="Send to node" ID="ID_1314687521"/>
+<node TEXT="Save script into node" ID="ID_920917811"/>
+</node>
+<node TEXT="r" ID="ID_906934789">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1563356019"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_981692423" LINK="#ID_971400958"/>
+</node>
+<node TEXT="Config + send to node" ID="ID_1377561715"/>
+<node TEXT="Select options and save script into node" ID="ID_923834124"/>
+</node>
+<node TEXT="r" ID="ID_319530479">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1904310838"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_893061771" LINK="#ID_27340568"/>
+</node>
+<node TEXT="Refresh Compiler Config" ID="ID_468100119"/>
+<node TEXT="Refreshes the Freeplane&apos;s **GroovyScript Compiler Configuration**&#xa;(libraries included)" ID="ID_1695144570"/>
+</node>
+<node TEXT="r" ID="ID_1058901683">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1047520427"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_885344875" LINK="#ID_603527740"/>
+</node>
+<node TEXT="Starts System.out interceptor" ID="ID_1164768265"/>
+<node TEXT="Restarts the ```systemOutInterceptor``` so that ```print()``` and ```println()``` get written in the log file or the console.&#xa;Use &apos;**View / Capture Standard Output**&apos; to direct the output to the console or to Freeplane&apos;s *log.0* file" ID="ID_823569848"/>
+</node>
+</node>
+</node>
+<node TEXT="Add-on configuration" ID="ID_1173958944">
+<node TEXT="local image" STYLE_REF="MarkdownHelperNode" ID="ID_1367297767"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/auto">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.imageLink(node,true)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="= edofro.MarkDownHelper.MDH.linkedNodeText(node)" STYLE_REF="MarkdownHelperLink" ID="ID_440247773" LINK="#ID_1628851653"/>
+</node>
+<node TEXT="You can configure some parameters of the add-on in Freeplane &apos;s preferences." ID="ID_1799291069"/>
+<node TEXT="table" STYLE_REF="MarkdownHelperNode" ID="ID_55445407"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
+<html>
+  <head>
+    
+  </head>
+  <body>
+    <p>
+      = edofro.MarkDownHelper.MDH.table(node)
+    </p>
+  </body>
+</html>
+</richcontent>
+<node TEXT="r" ID="ID_787350962">
+<node TEXT="parameter" ID="ID_309855458"/>
+<node TEXT="Description" ID="ID_738671987"/>
+<node TEXT="Default value" ID="ID_582708346"/>
+<node TEXT="Notes" ID="ID_26177819"/>
+</node>
+<node TEXT="r" ID="ID_263184019">
+<node TEXT="Attribute for file extensions" ID="ID_616708116"/>
+<node TEXT="Name of the attribute to indicate the file extension associated with the node" ID="ID_1455090415"/>
+<node TEXT="file_ext" ID="ID_238409262"/>
+<node TEXT="When you decide to save a script as a node&apos;s note, it will &apos;mark&apos; this node by defining its details to &apos;.groovy&apos;. But if its details have a text allready, then it will add a new attribute to the node to do this. The name of this attribute is defined in this preference." ID="ID_886190508"/>
+</node>
+<node TEXT="r" ID="ID_714566041">
+<node TEXT="Open in full Screen mode?" ID="ID_554660444"/>
+<node TEXT="Groovy Console must open in full screen mode" ID="ID_1282776616"/>
+<node TEXT="true" ID="ID_1183067633"/>
+<node TEXT="I choose true because I work with two monitors and this way it opens allways in fullscreen mode in the other monitor." ID="ID_1996770372"/>
+</node>
+<node TEXT="r" ID="ID_1485106552">
+<node TEXT="Show console button labels" ID="ID_1537521366"/>
+<node TEXT="Shows the labels of the buttons added to thr Groovey Console&apos;s toolbar" ID="ID_469289346"/>
+<node TEXT="true" ID="ID_1385463321"/>
+<node TEXT="The default value is true, so new users can become familiar with this buttons.&#xa;I have this option turned off" ID="ID_1907512571"/>
+</node>
+<node TEXT="r" ID="ID_1916174884">
+<node TEXT="Default writing target" ID="ID_1050402361"/>
+<node TEXT="Defines which is your preferred destination when writing (saving) a script into a node:&#xa;  0. its note&#xa;  1. &quot;script1&quot; attribute" ID="ID_904850384"/>
+<node TEXT="1. &quot;script1&quot; attribute" ID="ID_1140313765"/>
+<node TEXT="Everytime you save a new script in a node, it asks you where do you want to save it in (node or script1).&#xa;This parameter indicates which is the preselected option.&#xa;Personally I prefer to save my scripts in the notes, because this way I can look at them easier, but using script1 has its advantages too: they can be directly executed by Freeplane." ID="ID_1106880070"/>
+</node>
+</node>
+</node>
+</node>
+<node TEXT="about wikdshell add-on" ID="ID_1300104306">
+<node TEXT="wikdshell is an add-on that opened a Groovy Console to write and use scripts with Freeplane." ID="ID_1272928470"/>
+<node TEXT="wikdshell works with Freeplane versions up to v1.9.15." ID="ID_1469717355"/>
+<node TEXT="But with the new versions starting as v1.10.1 it doesn&apos;t work any longer." ID="ID_1729938505"/>
+</node>
+</node>
+</node>
+<node TEXT="README.MD" ID="ID_181733889" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/README.MD"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+</node>
+<node TEXT="resources" ID="ID_1818290900" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/">
+<attribute NAME="lastModifiedTime" VALUE="17-04-22 10:42" OBJECT="org.freeplane.features.format.FormattedDate|2022-04-17T10:42-0400|datetime"/>
+<attribute NAME="lastAccessTime" VALUE="25-04-22 10:04" OBJECT="org.freeplane.features.format.FormattedDate|2022-04-25T10:04-0400|dd-MM-yy HH:mm"/>
+<attribute NAME="creationTime" VALUE="17-04-22 10:42" OBJECT="org.freeplane.features.format.FormattedDate|2022-04-17T10:42-0400|datetime"/>
+<attribute NAME="fileSize" VALUE="0" OBJECT="org.freeplane.features.format.FormattedNumber|0|#,##0"/>
+<node TEXT="groovyConsole_001.png" ID="ID_1941588475" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_001.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+</node>
+<node TEXT="groovyConsole_002.png" ID="ID_1984441739" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_002.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+</node>
+<node TEXT="groovyConsole_003.png" ID="ID_1034407721" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_003.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+</node>
+<node TEXT="groovyConsole_004.png" ID="ID_1693528645" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_004.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+</node>
+<node TEXT="groovyConsole_005.png" ID="ID_1736357139" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_005.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+<richcontent CONTENT-TYPE="plain/markdown" TYPE="DETAILS"/>
+</node>
+<node TEXT="groovyConsole_006.png" ID="ID_971400958" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_006.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+<richcontent CONTENT-TYPE="plain/markdown" TYPE="DETAILS"/>
+</node>
+<node TEXT="groovyConsole_007.png" ID="ID_27340568" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_007.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+<richcontent CONTENT-TYPE="plain/markdown" TYPE="DETAILS"/>
+</node>
+<node TEXT="groovyConsole_008.png" ID="ID_603527740" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_008.png"><richcontent TYPE="NOTE" CONTENT-TYPE="plain/markdown"/>
+<richcontent CONTENT-TYPE="plain/markdown" TYPE="DETAILS"/>
+</node>
+<node TEXT="groovyConsole_009.png" ID="ID_1628851653" LINK="file:/C:/Users/Edo/Documents/GitHub/Freeplane_groovyConsole/resources/groovyConsole_009.png"/>
+</node>
 </node>
 </node>
 </map>